Renaming and removing enhanced image
