--- components/common/EnhancedImage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './EnhancedImage'
