+++ components/common/Footer/Footer.tsx
@@ -6,7 +6,7 @@ import type { Page } from '@bigcommerce/storefront-data-hooks/api/operations/get
 import getSlug from '@lib/get-slug'
 import { Github } from '@components/icons'
 import { Logo, Container } from '@components/ui'
-import { I18nWidget } from '@components/core'
+import { I18nWidget } from '@components/common'
 import s from './Footer.module.css'
 
 interface Props {
