+++ components/common/Navbar/Navbar.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 import Link from 'next/link'
 import s from './Navbar.module.css'
 import { Logo } from '@components/ui'
-import { Searchbar, UserNav } from '@components/core'
+import { Searchbar, UserNav } from '@components/common'
 interface Props {
   className?: string
 }
