+++ components/common/index.ts
@@ -9,4 +9,3 @@ export { default as Toggle } from './Toggle'
 export { default as Head } from './Head'
 export { default as HTMLContent } from './HTMLContent'
 export { default as I18nWidget } from './I18nWidget'
-export { default as EnhancedImage } from './EnhancedImage'
