+++ pages/[...pages].tsx
@@ -8,7 +8,7 @@ import getPage from '@bigcommerce/storefront-data-hooks/api/operations/get-page'
 import getAllPages from '@bigcommerce/storefront-data-hooks/api/operations/get-all-pages'
 import getSlug from '@lib/get-slug'
 import { missingLocaleInPages } from '@lib/usage-warns'
-import { Layout, HTMLContent } from '@components/core'
+import { Layout, HTMLContent } from '@components/common'
 
 export async function getStaticProps({
   preview,
