+++ pages/profile.tsx
@@ -2,7 +2,7 @@ import type { GetStaticPropsContext } from 'next'
 import { getConfig } from '@bigcommerce/storefront-data-hooks/api'
 import getAllPages from '@bigcommerce/storefront-data-hooks/api/operations/get-all-pages'
 import useCustomer from '@bigcommerce/storefront-data-hooks/use-customer'
-import { Layout } from '@components/core'
+import { Layout } from '@components/common'
 import { Container, Text } from '@components/ui'
 
 export async function getStaticProps({
