+++ pages/wishlist.tsx
@@ -2,7 +2,7 @@ import type { GetStaticPropsContext } from 'next'
 import { getConfig } from '@bigcommerce/storefront-data-hooks/api'
 import getAllPages from '@bigcommerce/storefront-data-hooks/api/operations/get-all-pages'
 import useWishlist from '@bigcommerce/storefront-data-hooks/wishlist/use-wishlist'
-import { Layout } from '@components/core'
+import { Layout } from '@components/common'
 import { Heart } from '@components/icons'
 import { Container, Text } from '@components/ui'
 import { WishlistCard } from '@components/wishlist'
